refactor(lfx-one): extract shared person-avatar + date-format utils
Eliminates triplicate copies of the avatar palette, initials hash, and
short-month date formatters across the three new People tabs. Each tab
now imports a single canonical helper from `@shared/utils/` instead of
carrying its own static-method + module-level-constant pair.

This also clears the personal `no-module-level-consts-or-local-
interfaces-in-apps-lfx-one` review pattern for these files — `SHORT_
MONTHS` is now an implementation detail of `date-format.util.ts` and
no longer appears in any component file.

- New `@shared/utils/person-avatar.util.ts` — `computePersonInitials`
  and `computePersonAvatarColorClass`, palette inlined as a
  module-private const
- New `@shared/utils/date-format.util.ts` — `formatMonthYearUtc`
  (Trainees enrolled/completed cells) and `formatLongDateUtc` (Event
  Attendees + Contributors expanded-row dates)
- Trainees, Event Attendees, Contributors: import + delegate;
  remove the static methods and the local SHORT_MONTHS / formatter

`all-employees.component.ts` keeps its own copies for now — that file
is not in this branch's scope; consolidating it is a follow-up.

Addresses: https://linuxfoundation.atlassian.net/browse/LFXV2-1874
Addresses: https://linuxfoundation.atlassian.net/browse/LFXV2-1875
Addresses: https://linuxfoundation.atlassian.net/browse/LFXV2-1876
